Inline postgres.js integration back into a single file
